v2.0.0 - Instant Object Creation Algorithm
The main algorithm used until this version was established to create an object list with all user data provided and to show/hide relevant objects based on the conditions matching the search filter.

With the "Instant Object Creation" algorithm, an object list will now be created for user data that matches the first letter of the search term of the current user, and then relevant objects will be shown/hidden based on the conditions matching the search filter. As a result, significantly fewer objects will be checked for each text input, and the script's performance will be greatly improved.